+++ src/dp/ungapped_simd.h
@@ -24,8 +24,8 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 namespace DP {
 
-DECL_DISPATCH(void, window_ungapped, (const Letter* query, const Letter** subjects, size_t subject_count, int window, int* out))
-DECL_DISPATCH(void, window_ungapped_best, (const Letter* query, const Letter** subjects, size_t subject_count, int window, int* out))
+DECL_DISPATCH(void, window_ungapped, (const Letter* query, const Letter** subjects, int subject_count, int window, int* out))
+DECL_DISPATCH(void, window_ungapped_best, (const Letter* query, const Letter** subjects, int subject_count, int window, int* out))
 
 }
